Added possibility to remove blog posts from the database.
Had some weird troubles trying to figure out why the blog list didn't rerender but as it appears it was rerendered. The function handling the sorting task just seemed to sort the blogs before waiting new ones to be fetched from the server. No all is good. The sorting function is now nested in the getAllBlogs() function where it is only called when the new blogs are fetched and the newBlogs is passed as an argument.