Refactoring is done for now. Notifications, Users and blogs are using redux state.
Bugs: When creating a blog, you don't see the delete button even though it's created by you. Refresh is needed. This has something to do with backend, that's not returning the full userinfo with the blog. More difficult with document database than relational database.

Still needing some refactoring, but the assignments for the course don't require ijt, so I'm choosing less work with this one and more work with other projects.